+++ all/build.gradle
@@ -23,7 +23,7 @@ def subprojects = [
         project(':opencensus-testing'),
         project(':opencensus-contrib-agent'),
         project(':opencensus-contrib-grpc-util'),
-        project(':opencensus-grpc-rpc-constants'),
+        project(':opencensus-contrib-grpc-metrics'),
         project(':opencensus-contrib-zpages'),
         project(':opencensus-exporter-trace-logging'),
         project(':opencensus-exporter-trace-stackdriver'),
@@ -34,7 +34,7 @@ def subprojects_javadoc = [
         project(':opencensus-api'),
         project(':opencensus-testing'),
         project(':opencensus-contrib-grpc-util'),
-        project(':opencensus-grpc-rpc-constants'),
+        project(':opencensus-contrib-grpc-metrics'),
         project(':opencensus-contrib-zpages'),
         project(':opencensus-exporter-trace-logging'),
         project(':opencensus-exporter-trace-stackdriver'),
