+++ contrib/grpc_metrics/README.md
@@ -1,3 +1,3 @@
-# OpenCensus gRPC RPC Constants
+# OpenCensus gRPC Metrics
 
 RPC measure and view constants used by gRPC.
