+++ settings.gradle
@@ -13,8 +13,8 @@ include ":core_impl"
 include ":core_impl_java"
 include ":core_impl_android"
 include ":opencensus-contrib-agent"
+include ":opencensus-contrib-grpc-metrics"
 include ":opencensus-contrib-grpc-util"
-include ":opencensus-grpc-rpc-constants"
 
 project(':opencensus-api').projectDir = "$rootDir/api" as File
 project(':opencensus-impl-core').projectDir = "$rootDir/impl_core" as File
@@ -22,8 +22,8 @@ project(':opencensus-impl-lite').projectDir = "$rootDir/impl_lite" as File
 project(':opencensus-impl').projectDir = "$rootDir/impl" as File
 project(':opencensus-testing').projectDir = "$rootDir/testing" as File
 project(':opencensus-contrib-agent').projectDir = "$rootDir/contrib/agent" as File
+project(':opencensus-contrib-grpc-metrics').projectDir = "$rootDir/contrib/grpc_metrics" as File
 project(':opencensus-contrib-grpc-util').projectDir = "$rootDir/contrib/grpc_util" as File
-project(':opencensus-grpc-rpc-constants').projectDir = "$rootDir/grpc/rpc_constants" as File
 project(':opencensus-exporter-trace-logging').projectDir = "$rootDir/exporters/trace_logging" as File
 project(':opencensus-exporter-trace-stackdriver').projectDir = "$rootDir/exporters/trace_stackdriver" as File
 project(':opencensus-exporter-trace-zipkin').projectDir = "$rootDir/exporters/trace_zipkin" as File
